+++ branches/openMP/src/initializer.cpp
@@ -209,10 +209,10 @@ bool InitGenotype::apply(Population & pop) const
 			for (; it.valid(); ++it)
 				for (vectoru::iterator p = ploidy.begin(); p != ploidy.end(); ++p)
 					for (vectoru::const_iterator loc = loci.begin(); loc != loci.end(); ++loc, ++idx)
-						it->setAllele(ToAllele(m_genotype[idx % sz]), *loc, *p);
+						it->setAllele(ToAllele(m_genotype[idx % sz]), *loc, static_cast<int>(*p));
 		} else if (!m_prop.empty()) {
 			WeightedSampler ws;
-			UINT sz = pop.subPopSize(*sp);
+			size_t sz = pop.subPopSize(*sp);
 			if (m_haplotypes.empty()) {
 				// initialize by allele. Gurantee proportion at each locus.
 				for (vectoru::const_iterator loc = loci.begin(); loc != loci.end(); ++loc) {
@@ -220,7 +220,7 @@ bool InitGenotype::apply(Population & pop) const
 					IndIterator it = pop.indIterator(sp->subPop());
 					for (; it.valid(); ++it)
 						for (vectoru::iterator p = ploidy.begin(); p != ploidy.end(); ++p)
-							it->setAllele(ToAllele(ws.draw()), *loc, *p);
+							it->setAllele(ToAllele(ws.draw()), *loc, static_cast<int>(*p));
 				}
 			} else {
 				ws.set(m_prop.begin(), m_prop.end(), sz * ploidy.size());
@@ -231,7 +231,7 @@ bool InitGenotype::apply(Population & pop) const
 						size_t hapSz = haplotype.size();
 						size_t j = 0;
 						for (vectoru::const_iterator loc = loci.begin(); loc != loci.end(); ++loc, ++j)
-							it->setAllele(ToAllele(haplotype[j % hapSz]), *loc, *p);
+							it->setAllele(ToAllele(haplotype[j % hapSz]), *loc, static_cast<int>(*p));
 					}
 			}
 		} else {
@@ -242,13 +242,13 @@ bool InitGenotype::apply(Population & pop) const
 				for (vectoru::iterator p = ploidy.begin(); p != ploidy.end(); ++p) {
 					if (m_haplotypes.empty()) {
 						for (vectoru::const_iterator loc = loci.begin(); loc != loci.end(); ++loc)
-							it->setAllele(ToAllele(ws.draw()), *loc, *p);
+							it->setAllele(ToAllele(ws.draw()), *loc, static_cast<int>(*p));
 					} else {
 						const vectori & haplotype = m_haplotypes[ws.draw()];
 						size_t hapSz = haplotype.size();
 						size_t j = 0;
 						for (vectoru::const_iterator loc = loci.begin(); loc != loci.end(); ++loc, ++j)
-							it->setAllele(ToAllele(haplotype[j % hapSz]), *loc, *p);
+							it->setAllele(ToAllele(haplotype[j % hapSz]), *loc, static_cast<int>(*p));
 					}
 				}
 		}
